all changes

+++ backend/routes/photoRoutes.js
@@ -1,21 +1,23 @@
 const express = require("express");
 const multer = require("multer");
-const Photo = require("../models/photo");
-const path = require("path");
+const { CloudinaryStorage } = require("multer-storage-cloudinary");
+const cloudinary = require("../config/cloudinary");
+const Photo = require("../models/Photo");
 
 const router = express.Router();
 
-// Storage config
-const storage = multer.diskStorage({
-  destination: "./uploads/",
-  filename: (req, file, cb) => {
-    cb(null, Date.now() + path.extname(file.originalname));
+// Cloudinary storage config
+const storage = new CloudinaryStorage({
+  cloudinary: cloudinary,
+  params: {
+    folder: "globe-photos",
+    allowed_formats: ["jpg", "png", "jpeg"],
   },
 });
 
 const upload = multer({ storage });
 
-// GET all photos
+// GET photos
 router.get("/", async (req, res) => {
   const photos = await Photo.find().sort({ createdAt: -1 });
   res.json(photos);
@@ -24,7 +26,8 @@ router.get("/", async (req, res) => {
 // Upload photo
 router.post("/", upload.single("image"), async (req, res) => {
   const photo = await Photo.create({
-    url: `http://localhost:5000/uploads/${req.file.filename}`,
+    url: req.file.path,       // Cloudinary URL
+    public_id: req.file.filename,
   });
 
   res.json(photo);
@@ -32,7 +35,13 @@ router.post("/", upload.single("image"), async (req, res) => {
 
 // Delete photo
 router.delete("/:id", async (req, res) => {
+  const photo = await Photo.findById(req.params.id);
+
+  if (!photo) return res.status(404).json({ message: "Not found" });
+
+  await cloudinary.uploader.destroy(photo.public_id);
   await Photo.findByIdAndDelete(req.params.id);
+
   res.json({ message: "Deleted" });
 });
